refactor: extract BrowserMixin from scraper.py
Move browser/MHTML methods (page, initialize_playwright, _capture_mhtml,
_fetch_soup_and_mhtml, etc.) into browser_mixin.py. BaseScraper now
inherits from BrowserMixin. Only 3 scrapers use browser (MA, PR, PE).
